palette

+++ src/MiNET/MiNET/Net/BedrockMessageHandlerBase.cs
@@ -156,6 +156,7 @@ namespace MiNET.Net
 				if (_session.InitializedCompression)
 				{
 					сompress = stream.ReadByte();
+					//Log.Debug(Packet.HexDump(payload));
 				}
 				try
 				{
@@ -221,7 +222,7 @@ namespace MiNET.Net
 								//if (Log.IsDebugEnabled)
 								//	Log.Debug($"0x{internalBuffer[0]:x2}\n{Packet.HexDump(internalBuffer)}");
 
-								messages.Add(PacketFactory.Create((byte) id, internalBuffer, "mcpe") ??
+								messages.Add(PacketFactory.Create((short) id, internalBuffer, "mcpe") ??
 											new UnknownPacket((byte) id, internalBuffer));
 							}
 							catch (Exception e)
